refactor(tel-ui): File organisation
