+++ telemetry/packages/ui/app/components/index.ts
@@ -1,6 +1,6 @@
-export { PodConnectionStatus } from './connection-status';
-export { Logo } from './logo';
-export { Latency } from './latency';
-export { PodControls } from './pod-controls';
-export { PodState } from './pod-state';
+export { PodConnectionStatus } from './sidebar/pod-connection-status';
+export { Logo } from './shared/logo';
+export { Latency } from './sidebar/latency';
+export { PodControls } from './sidebar/pod-controls';
+export { PodState } from './shared/pod-state';
 export { PodDisconnectError } from './pod-disconnect-error';
