+++ telemetry/packages/ui/app/components/shared/logo.tsx
