+++ telemetry/packages/ui/app/components/shared/pod-state.tsx
@@ -6,7 +6,7 @@ import {
   NULL_STATES,
   ACTIVE_STATES,
 } from '@hyped/telemetry-constants';
-import { Button } from './ui/button';
+import { Button } from '../ui/button';
 import {
   Dialog,
   DialogContent,
@@ -15,7 +15,7 @@ import {
   DialogTitle,
   DialogTrigger,
 } from '@/components/ui/dialog';
-import { StateMachineFlowChart } from './flow/flow-chart';
+import { StateMachineFlowChart } from '../views/state-machine/flow-chart';
 
 /**
  * Displays the current state of the pod.
