+++ telemetry/packages/ui/app/components/sidebar/pod-connection-status.tsx
