+++ telemetry/packages/ui/app/components/sidebar/pod-selector.tsx
@@ -1,12 +1,12 @@
 import { POD_IDS, PodId, pods } from '@hyped/telemetry-constants';
-import { Label } from './ui/label';
+import { Label } from '../ui/label';
 import {
   Select,
   SelectContent,
   SelectItem,
   SelectTrigger,
   SelectValue,
-} from './ui/select';
+} from '../ui/select';
 import { useCurrentPod } from '@/context/pods';
 
 export const PodSelector = () => {
