+++ telemetry/packages/ui/app/components/views/debug-view/debug-view.tsx
