+++ telemetry/packages/ui/app/components/views/debug-view/pod-state.tsx
@@ -12,7 +12,7 @@ import {
   CardDescription,
   CardHeader,
   CardTitle,
-} from '../ui/card';
+} from '@/components/ui/card';
 import { CircleDashed } from 'lucide-react';
 
 const styles = {
