+++ telemetry/packages/ui/app/components/views/state-machine/flow-chart.tsx
