+++ telemetry/packages/ui/app/components/views/state-machine/nodes/active-node.tsx
