+++ telemetry/packages/ui/app/components/views/state-machine/nodes/failure-node.tsx
