+++ telemetry/packages/ui/app/components/views/state-machine/nodes/passive-node.tsx
