+++ telemetry/packages/ui/app/components/views/state-machine/nodes/styles.ts
