+++ telemetry/packages/ui/app/components/views/state-machine/nodes/text-node.tsx
