+++ telemetry/packages/ui/app/components/views/state-machine/state-machine.tsx
