+++ telemetry/packages/ui/app/components/views/state-machine/styles.css
