+++ telemetry/packages/ui/app/components/views/state-machine/types.ts
