+++ telemetry/packages/ui/app/components/views/state-machine/utils.ts
